refactor specs
there is a condition where the specs are flaky because the with_model has had the .not_scoped
set and for some reason it is sticky-- even with another model defined.  Rather than wrestle
with it, easier to just break the tests out...
