+++ spec/soft_delete/restorable_spec.rb
@@ -18,13 +18,12 @@ RSpec.describe SoftDelete::Restorable do
     end
   end
 
-  describe ".deleted" do
-    context "when there is more than one default scope" do
-
+  describe '.deleted' do
+    context 'when there is more than one default scope' do
       let!(:without_title) { Note.create(body: 'blah blah', deleted_at: Time.now) }
       let!(:with_title) { Note.create(body: 'blah blah', title: 'some title', deleted_at: Time.now) }
 
-      it "only removes the default scope associated with deleted_at" do
+      it 'only removes the default scope associated with deleted_at' do
         expect(Note.deleted.count).to eq(1)
         expect(Note.deleted.first).to eq(with_title)
       end
@@ -34,20 +33,20 @@ RSpec.describe SoftDelete::Restorable do
   describe '#restore_soft_delete' do
     subject { note.restore_soft_delete(validate: validate) }
 
-    let(:note) { Note.create(body: 'blah blah', title: "some note", deleted_at: Time.now) }
+    let(:note) { Note.create(body: 'blah blah', title: 'some note', deleted_at: Time.now) }
     let(:validate) { true }
 
-    it "restores the record" do
-      expect { subject }.to change{ Note.count }.by(1)
+    it 'restores the record' do
+      expect { subject }.to change { Note.count }.by(1)
     end
 
-    context "when skipping validations" do
-      let(:note) { Note.create(title: "some note", deleted_at: Time.now) }
+    context 'when skipping validations' do
+      let(:note) { Note.create(title: 'some note', deleted_at: Time.now) }
       let(:validate) { false }
 
-      it "restores an otherwise invalid record" do
+      it 'restores an otherwise invalid record' do
         expect(note.valid?).to eq(false)
-        expect { subject }.to change{ Note.count }.by(1)
+        expect { subject }.to change { Note.count }.by(1)
       end
     end
   end
