+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
-require "bundler/setup"
-require "soft_delete"
+require 'bundler/setup'
+require 'soft_delete'
 
 require 'database_cleaner'
 require 'with_model'
@@ -22,7 +22,7 @@ RSpec.configure do |config|
   config.extend WithModel
 
   # Enable flags like --only-failures and --next-failure
-  config.example_status_persistence_file_path = ".rspec_status"
+  config.example_status_persistence_file_path = '.rspec_status'
 
   # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!
